1.6.17 (#5)
* pnglibconf.h: define PNG_ZLIB_VERNUM 0 (unknown)

With autotools then pnglibconf.h is generated using awk and the version
of zlib found at configure time gets captured by defining PNG_ZLIB_VERNUM
to whatever ZLIB_VERNUM is defined as. This enables a compile-time
assertion that this configuration of libpng is then only ever used with
that same version of zlib.

If building libng with cmake then they avoid depending on this awk
script and instead leave PNG_ZLIB_VERNUM defined as zero, which means
'unknown' and any version of zlib can be used (assuming the api is
otherwise compatible with what libpng expects).

While avoiding the complication of running the awk script we also don't
want to be tied to a specific version of zlib. Like cmake we use a copy
of scripts/pnglibconf.h.prebuilt for our pnglibconf.h, with
PNG_ZLIB_VERNUM defined as zero.

* Add missing -lm dependency